+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 23.06.1
 * Added app version metric to every request sent.
-* Fixed a bug that caused build issues.
+* Fixed a bug that caused build issues when not running inside the editor.
 
 ## 23.06.0
 * Deprecated "CrushReports" and "OptionalParameters" getters, in "Countly" class are removed.